requirements.txt
python-telegram-bot==21.0
python-dotenv==1.0.0
pandas==2.0.3
openpyxl==3.1.2
XlsxWriter==3.1.2
Flask==2.3.2
pytz==2023.3
APScheduler==3.10.1
telegram